fix(pipeline): satisfy the file-size ratchet — extract the judge stage into its own module
CI's ratchet flagged pipeline.rs growing 35 lines past its grandfathered
ceiling. Per the guard's own prescription ('put the new code in its own
module instead'), the // Stage: judge section — judge(), judge_guidance(),
and warn_judge_caveat() — moves to pipeline/judge_stage.rs, the same split
witness_stage.rs and disclosure.rs already took. pipeline.rs drops to
3,296 lines, 85 under its ceiling; no behavior changes.

The ratchet also fails on the BASE branch: #1202 grew stella-model/src/zai.rs
to 1,524 lines against its 1,520 baseline ceiling without the baseline diff.
This bumps that one entry (the guard's sanctioned visible-diff path for
grandfathered growth) so this PR's gate can go green at all.
