+++ scripts/file-size-baseline.txt
@@ -25,7 +25,7 @@
 1849 stella-fleet/src/fleet.rs
 1628 stella-model/src/anthropic/tests.rs
 2060 stella-model/src/openai.rs
-1520 stella-model/src/zai.rs
+1524 stella-model/src/zai.rs
 1773 stella-model/src/zai/tests.rs
 3381 stella-pipeline/src/pipeline.rs
 2440 stella-pipeline/src/pipeline/tests.rs
